LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/106-module-noimport-expected.txt

FAILED (This TC requires JavaScript enabled)

FAIL  scheduler: stylesheets blocking module scripts assert_equals: expected "fixed" but got "static"
Test
